Add files via upload
Successfully created a for loop for Nike to cycle through and get 700 shoes from the men's shoes page. Currently working on how to properly convert the dictionary object "output" to a txt file where a new line is started at 'cloudProductId': and ends after 'minimumAdvertisedPrice':. Also need to fix issue where 'cloudProductId' is repeated for the same shoe, i.e. that info is in two different print statements. Take a closer look at all the info that's created when you pull from the website.